Add logo and component header and header.css + add import for the App page
