+++ client/src/App.tsx
@@ -1,6 +1,7 @@
 import { Outlet } from "react-router-dom";
 import "./App.css";
-import Header from "./components/Header";
+import Header from "./components/Header/Header";
+import "./style/globals.css";
 
 function App() {
   return (
