--- client/src/components/Header.tsx
@@ -1,17 +1,0 @@
-function Header() {
-  return (
-    <>
-      <nav className="header-container-link">
-        <img src="/src/assets/images/Logo.png" alt="" />
-        <ul className="">
-          <li>Accueil</li>
-          <li>Société</li>
-          <li>Environnement</li>
-          <li>Contact</li>
-        </ul>
-      </nav>
-    </>
-  );
-}
-
-export default Header;
